vkd3d: Fix formatting warnings.
Extremely annoying and there's a reason why we'd had -Wno-format,
but there were tons of (mostly benign) formatting bugs all over the place.

Native and MinGW builds use different types for things mostly due to
long. There seems to be no good way to have MinGW assume that %zu works since it still
targets msvcrt.dll from prehistoric era, so needed a hack to get around that too.

+++ include/private/vkd3d_native_sync_handle.h
@@ -72,7 +72,7 @@ static inline int vkd3d_native_sync_handle_release(vkd3d_native_sync_handle hand
             /* Failing to release semaphore is expected if the counter exceeds the maximum limit.
              * If the application does not wait for the semaphore once per present, this
              * will eventually happen. */
-            TRACE("Failed to release semaphore (#%x).\n", GetLastError());
+            TRACE("Failed to release semaphore (#%x).\n", (int)GetLastError());
             prev = -1;
         }
         return prev;
@@ -81,7 +81,7 @@ static inline int vkd3d_native_sync_handle_release(vkd3d_native_sync_handle hand
     {
         if (!SetEvent(handle.handle))
         {
-            ERR("Failed to set event %p (#%x).\n", handle.handle, GetLastError());
+            ERR("Failed to set event %p (#%x).\n", handle.handle, (int)GetLastError());
             return -1;
         }
         return 0;
